requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Google OAuth and Gmail API
google-auth-oauthlib==1.0.0
google-api-python-client==2.103.0
google-auth==2.23.3


# Google Generative AI (Gemini)
google-generativeai==0.3.2

# Standard library dependencies (included for completeness, but usually pre-installed)
# These are typically part of Python standard library but listing for clarity:
# - base64 (built-in)
# - datetime (built-in) 
# - re (built-in)
# - html (built-in)
# - os (built-in)
# - threading (built-in)
# - time (built-in)
# - uuid (built-in)
# - pickle (built-in)
# - json (built-in)
# - random (built-in)
# - asyncio (built-in)
# - concurrent.futures (built-in) 